Move emp_frequency tests to phyloInfo
Moved the tests for computing empirical frequencies to phyloInfo instead
of the likelihood tests.
