+++ phylo/src/likelihood/likelihood_tests.rs
@@ -11,12 +11,11 @@ use crate::evolutionary_models::{
     EvolutionaryModel,
     ProteinModelType::{self, *},
 };
-use crate::frequencies;
 use crate::likelihood::LikelihoodCostFunction;
 use crate::phylo_info::{PhyloInfo, PhyloInfoBuilder};
 use crate::substitution_models::dna_models::DNASubstModel;
 use crate::substitution_models::protein_models::{ProteinLikelihoodCost, ProteinSubstModel};
-use crate::substitution_models::{FreqVector, SubstMatrix, SubstitutionLikelihoodCost};
+use crate::substitution_models::{SubstMatrix, SubstitutionLikelihoodCost};
 use crate::tree::{tree_parser, NodeIdx::Leaf as L, Tree};
 
 #[cfg(test)]
@@ -379,77 +378,5 @@ fn huelsenbeck_example_dna_reversibility_likelihood(
         LikelihoodCostFunction::compute_logl(&likelihood),
         LikelihoodCostFunction::compute_logl(&likelihood_rerooted),
         epsilon = 1e-10,
-    );
-}
-
-#[test]
-fn empirical_frequencies_no_ambigs() {
-    let sequences = Sequences::new(vec![
-        Record::with_attrs("one", None, b"CCCCCCCC"),
-        Record::with_attrs("two", None, b"AAAAAAAA"),
-        Record::with_attrs("three", None, b"TTTTTTTT"),
-        Record::with_attrs("four", None, b"GGGGGGGG"),
-    ]);
-    let newick = "((one:2,two:2):1,(three:1,four:1):2);".to_string();
-    let info = PhyloInfoBuilder::build_from_objects(sequences, tree_newick(&newick)).unwrap();
-    assert_relative_eq!(info.freqs(), frequencies!(&[0.25; 4]), epsilon = 1e-6);
-}
-
-#[test]
-fn empirical_frequencies_ambig_x() {
-    let sequences = Sequences::new(vec![
-        Record::with_attrs("on", None, b"XXXXXXXX"),
-        Record::with_attrs("tw", None, b"XXXXXXXX"),
-        Record::with_attrs("th", None, b"NNNNNNNN"),
-        Record::with_attrs("fo", None, b"NNNNNNNN"),
-    ]);
-    let newick = "((on:2,tw:2):1,(th:1,fo:1):2);".to_string();
-    let info = PhyloInfoBuilder::build_from_objects(sequences, tree_newick(&newick)).unwrap();
-    assert_relative_eq!(info.freqs(), frequencies!(&[0.25; 4]), epsilon = 1e-6);
-}
-
-#[test]
-fn empirical_frequencies_ambig_n() {
-    let sequences = Sequences::new(vec![
-        Record::with_attrs("on", None, b"AAAAAAAAAA"),
-        Record::with_attrs("tw", None, b"XXXXXXXXXX"),
-        Record::with_attrs("th", None, b"CCCCCCCCCC"),
-        Record::with_attrs("fo", None, b"NNNNNNNNNN"),
-    ]);
-    let newick = "(((on:2,tw:2):1,th:1):4,fo:1);".to_string();
-    let info = PhyloInfoBuilder::build_from_objects(sequences, tree_newick(&newick)).unwrap();
-    assert_relative_eq!(
-        info.freqs(),
-        frequencies!(&[0.125, 0.375, 0.375, 0.125]),
-        epsilon = 1e-6
-    );
-}
-
-#[test]
-fn empirical_frequencies_ambig_other() {
-    let sequences = Sequences::new(vec![
-        Record::with_attrs("A", None, b"VVVVVVVVVV"),
-        Record::with_attrs("B", None, b"TTTTVVVTVV"),
-    ]);
-    let newick = "(A:2,B:2):1.0;".to_string();
-    let info = PhyloInfoBuilder::build_from_objects(sequences, tree_newick(&newick)).unwrap();
-    assert_relative_eq!(info.freqs(), frequencies!(&[0.25; 4]), epsilon = 1e-6);
-    let sequences = Sequences::new(vec![
-        Record::with_attrs("A", None, b"SSSSSSSSSSSSSSSSSSSS"),
-        Record::with_attrs("B", None, b"WWWWWWWWWWWWWWWWWWWW"),
-    ]);
-    let info = PhyloInfoBuilder::build_from_objects(sequences, tree_newick(&newick)).unwrap();
-    assert_relative_eq!(info.freqs(), frequencies!(&[0.25; 4]), epsilon = 1e-6);
-}
-
-#[test]
-fn empirical_frequencies_no_aas() {
-    let sequences = Sequences::new(vec![Record::with_attrs("A", None, b"BBBBBBBBB")]);
-    let newick = "A:1.0;".to_string();
-    let info = PhyloInfoBuilder::build_from_objects(sequences, tree_newick(&newick)).unwrap();
-    assert_relative_eq!(
-        info.freqs(),
-        frequencies!(&[3.0 / 10.0, 3.0 / 10.0, 1.0 / 10.0, 3.0 / 10.0]),
-        epsilon = 1e-6
     );
 }
